+++ xclim/atmos/_precip.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-import xclim.indices.threshold
-from xclim.indices import indices as _ind
+
+from xclim import indices
 from xclim.utils import Indicator, Indicator2D
 
 
@@ -23,7 +23,7 @@ rain_on_frozen_ground_days = PrTas(identifier='rain_frzgr',
                                                "Precipitation is assumed to be rain when the"
                                                "daily average temperature is above 0℃.",
                                    cell_methods='',
-                                   compute=_ind.rain_on_frozen_ground_days,
+                                   compute=indices.rain_on_frozen_ground_days,
                                    )
 
 max_1day_precipitation_amount = Pr(identifier='rx1day',
@@ -32,7 +32,7 @@ max_1day_precipitation_amount = Pr(identifier='rx1day',
                                    long_name='maximum 1-day total precipitation',
                                    description="{freq} maximum 1-day total precipitation",
                                    cellmethods='time: sum within days time: maximum over days',
-                                   compute=_ind.max_1day_precipitation_amount,
+                                   compute=indices.max_1day_precipitation_amount,
                                    )
 
 max_n_day_precipitation_amount = Pr(identifier='rx{window}day',
@@ -41,7 +41,7 @@ max_n_day_precipitation_amount = Pr(identifier='rx{window}day',
                                     long_name='maximum {window}-day total precipitation',
                                     description="{freq} maximum {window}-day total precipitation",
                                     cellmethods='time: sum within days time: maximum over days',
-                                    compute=_ind.max_n_day_precipitation_amount,
+                                    compute=indices.max_n_day_precipitation_amount,
                                     )
 
 wetdays = Pr(identifier='r{thresh}mm',
@@ -50,7 +50,7 @@ wetdays = Pr(identifier='r{thresh}mm',
              long_name='Number of Wet Days (precip >= {thresh} mm)',
              description='{freq} number of days with daily precipitation over {thresh} mm',
              cell_methods='time: sum within days time: sum over days',
-             compute=xclim.indices.threshold.wetdays,
+             compute=indices.wetdays,
              )
 
 maximum_consecutive_wet_days = Pr(identifier='cwd',
@@ -61,7 +61,7 @@ maximum_consecutive_wet_days = Pr(identifier='cwd',
                                   description='{freq} maximum number of days with daily '
                                               'precipitation over {thresh} mm',
                                   cell_methods='time: sum within days time: sum over days',
-                                  compute=xclim.indices.threshold.maximum_consecutive_wet_days,
+                                  compute=indices.maximum_consecutive_wet_days,
                                   )
 
 maximum_consecutive_dry_days = Pr(identifier='cdd',
@@ -72,7 +72,7 @@ maximum_consecutive_dry_days = Pr(identifier='cdd',
                                   description='{freq} maximum number of days with daily '
                                               'precipitation below {thresh} mm',
                                   cell_methods='time: sum within days time: sum over days',
-                                  compute=_ind.maximum_consecutive_dry_days,
+                                  compute=indices.maximum_consecutive_dry_days,
                                   )
 
 daily_pr_intensity = Pr(identifier='sdii',
@@ -82,7 +82,7 @@ daily_pr_intensity = Pr(identifier='sdii',
                         description="{freq} Simple Daily Intensity Index (SDII) : {freq} average precipitation "
                                     "for days with daily precipitation over {thresh} mm",
                         cell_methods='',
-                        compute=xclim.indices.threshold.daily_pr_intensity,
+                        compute=indices.daily_pr_intensity,
                         )
 
 precip_accumulation = Pr(identifier='prcptot',
@@ -91,5 +91,5 @@ precip_accumulation = Pr(identifier='prcptot',
                          long_name='Total precipitation',
                          description='{freq} total precipitation',
                          cell_methods='time: sum within days time: sum over days',
-                         compute=_ind.precip_accumulation
+                         compute=indices.precip_accumulation
                          )
